+++ src/main/java/seedu/duke/storage/banks/FoodBankStorage.java
@@ -1,7 +1,7 @@
 package seedu.duke.storage.banks;
 
 import seedu.duke.item.ItemBank;
-import seedu.duke.storage.StorageNew;
+import seedu.duke.storage.Storage;
 import seedu.duke.storage.exceptions.UnableToReadFileException;
 import seedu.duke.storage.exceptions.UnableToWriteFileException;
 
@@ -9,7 +9,7 @@ import java.io.FileNotFoundException;
 import java.util.ArrayList;
 import java.util.logging.Level;
 
-public class FoodBankStorage extends StorageNew {
+public class FoodBankStorage extends Storage {
 
     public static final String FILENAME_BANK_FOOD = "food_bank.txt";
     public static final String FILEPATH_BANK_FOOD = FILEPATH + FILENAME_BANK_FOOD;
